# Add benchmarks and optimize config/build targets