--- a/tests/benchmarks/build_targets/.gitignore
+++ b/tests/benchmarks/build_targets/.gitignore
@@ -1,0 +1,8 @@
+# Xmake cache
+.xmake/
+build/
+
+# MacOS Cache
+.DS_Store
+
+