--- a/tests/benchmarks/build_targets/meson.build
+++ b/tests/benchmarks/build_targets/meson.build
@@ -1,5 +1,5 @@
 project('test', 'cpp')
 
 foreach i : range(1, 31)
-    executable('test' + i.to_string(), 'src/main.cpp')
+    executable('test' + i.to_string(), 'src/test_' + i.to_string() + '.cpp')
 endforeach