--- a/tests/benchmarks/build_targets/src/test_23.cpp
+++ b/tests/benchmarks/build_targets/src/test_23.cpp
@@ -1,0 +1,6 @@
+#include <iostream>
+
+int main(int argc, char** argv) {
+    std::cout << "hello world!" << std::endl;
+    return 0;
+}