--- a/tests/benchmarks/build_targets/test.lua
+++ b/tests/benchmarks/build_targets/test.lua
@@ -3,6 +3,10 @@ import("core.tool.toolchain")
 
 function test_build(t)
 
+    if xmake.is_embed() then
+        return
+    end
+
     -- xmake
     os.tryrm("build")
     local xmake_dt = os.mclock()