--- a/tests/benchmarks/build_targets/xmake.lua
+++ b/tests/benchmarks/build_targets/xmake.lua
@@ -1,0 +1,7 @@
+add_rules("mode.debug", "mode.release")
+
+for i = 1, 30 do
+    target("test" .. i)
+        set_kind("binary")
+        add_files("src/main.cpp")
+end