--- a/tests/benchmarks/config_targets/.gitignore
+++ b/tests/benchmarks/config_targets/.gitignore
@@ -1,0 +1,8 @@
+# Xmake cache
+.xmake/
+build/
+
+# MacOS Cache
+.DS_Store
+
+