--- a/tests/benchmarks/config_targets/meson.build
+++ b/tests/benchmarks/config_targets/meson.build
@@ -1,0 +1,5 @@
+project('test', 'cpp')
+
+foreach i : range(1, 1001)
+    executable('test' + i.to_string(), 'src/main.cpp')
+endforeach