--- a/xmake/core/sandbox/modules/import/core/tool/toolchain.lua
+++ b/xmake/core/sandbox/modules/import/core/tool/toolchain.lua
@@ -30,6 +30,7 @@ local raise     = require("sandbox/modules/raise")
 -- inherit some builtin interfaces
 sandbox_core_tool_toolchain.apis        = toolchain.apis
 sandbox_core_tool_toolchain.directories = toolchain.directories
+sandbox_core_tool_toolchain.save        = toolchain.save
 
 -- get all toolchains list
 function sandbox_core_tool_toolchain.list()