--- a/xmake/plugins/project/vsxmake/getinfo.lua
+++ b/xmake/plugins/project/vsxmake/getinfo.lua
@@ -22,6 +22,7 @@
 import("core.base.option")
 import("core.base.semver")
 import("core.base.hashset")
+import("core.tool.toolchain")
 import("core.project.config")
 import("core.project.project")
 import("core.platform.platform")
@@ -483,6 +484,9 @@ function main(outputdir, vsinfo)
 
             -- update config files
             generate_configfiles()
+
+            -- save toolchain configs
+            toolchain.save()
 
             -- ensure to enter project directory
             os.cd(project.directory())