--- a/xmake/toolchains/c51/xmake.lua
+++ b/xmake/toolchains/c51/xmake.lua
@@ -36,7 +36,6 @@ toolchain("c51")
         local c51 = find_c51()
         if c51 and c51.sdkdir and find_tool("c51") then
             toolchain:config_set("sdkdir", c51.sdkdir)
-            toolchain:configs_save()
             return true
         end
     end)