# Config targets

## macOS

### Before optimization

```console
ruki:xmake ruki$ xrepo env -b "cmake,ninja,meson" xmake l ./tests/run.lua config_targets
> 1 test(s) found
>> [1/1]: testing tests/benchmarks/config_targets ...
config targets/1k: xmake: 2372 ms
config targets/1k: cmake/ninja: 2187 ms
config targets/1k: cmake/makefile: 6229 ms
config targets/1k: meson: 3484 ms
> 1 test(s) succeed
```

### After optimization

```console
ruki:xmake ruki$ xrepo env -b "cmake,ninja,meson" xmake l ./tests/run.lua config_targets
> 1 test(s) found
>> [1/1]: testing tests/benchmarks/config_targets ...
config targets/1k: xmake: 2159 ms
config targets/1k: cmake/default: 5883 ms
config targets/1k: cmake/ninja: 2222 ms
config targets/1k: cmake/makefile: 6097 ms
config targets/1k: meson: 3524 ms
> 1 test(s) succeed
```

## Windows

### Before optimization

```console
PS D:\projects\personal\xmake> xrepo env -b "msvc,cmake,ninja,meson" xmake l .\tests\run.lua config_targets
> 1 test(s) found
>> [1/1]: testing tests\benchmarks\config_targets ...
config targets/1k: xmake:  <-------------------------------------- stuck
> 1 test(s) succeed
```

### After optimization

```console
PS D:\projects\personal\xmake> xrepo env -b "msvc,cmake,ninja,meson" xmake l .\tests\run.lua config_targets
> 1 test(s) found
>> [1/1]: testing tests\benchmarks\config_targets ...
config targets/1k: xmake: 4047 ms
config targets/1k: cmake/default: 22907 ms
config targets/1k: cmake/ninja: 5859 ms
config targets/1k: meson: 7469 ms
> 1 test(s) succeed
```

https://github.com/xmake-io/xmake/issues/6775