@waruqi 2025-09-07
# Build targets

## macOS

```console
ruki:xmake ruki$ xrepo env -b "cmake,ninja,meson" xmake l ./tests/run.lua build_targets
> 1 test(s) found
>> [1/1]: testing tests/benchmarks/build_targets ...
build targets/30: xmake: 4690 ms
build targets/30: cmake/default: 31461 ms
build targets/30: cmake/ninja: 7490 ms
build targets/30: cmake/makefile: 32648 ms
build targets/30: meson: 16120 ms
> 1 test(s) succeed
```

## Windows

```console
PS D:\projects\personal\xmake> xrepo env -b "msvc,cmake,ninja,meson" xmake l .\tests\run.lua build_targets
> 1 test(s) found
>> [1/1]: testing tests\benchmarks\build_targets ...
build targets/30: xmake: 7875 ms
build targets/30: cmake/default: 77016 ms
build targets/30: cmake/ninja: 15594 ms
build targets/30: meson: 15016 ms
> 1 test(s) succeed
```